+++ dosee-functions.js
@@ -49,14 +49,14 @@
 
     // Extracts the URL query string and run it through the URLSearchParams API
     DOSee.newQueryString = () => {
-        const wlh = window.location.href
-        const start = wlh.indexOf(`?`)
-        const end = wlh.indexOf(`#`)
+        const url = window.location.href
+        const start = url.indexOf(`?`)
+        const end = url.indexOf(`#`)
         // URLSearchParams API works best with only the URL query string rather than the whole URL
         // slice string value between ? and #
-        if (end >= 0) return new URLSearchParams(wlh.slice(start, end))
+        if (end >= 0) return new URLSearchParams(url.slice(start, end))
         // slice string value starting at ? to the end of the URL
-        else return new URLSearchParams(wlh.slice(start))
+        else return new URLSearchParams(url.slice(start))
     }
 
     // Updates the content data stored in a HTML <meta> tag